Merged PR 866815: Replace nodetool with novatool
Medusa runs nodetool command to create snapshot.  Because we switch to Mariner, if we want to execute nodetool commands, we need to add "-i" before "nodetool" and the command will be "sudo -i nodetool ... ". Otherwise, it will throw Nodetool class not found error.

Since our nova provides novatool command, which wraps nodetool command with some necessary parameters like jmx password, Medusa could use novatool directly. If we change the parameters of nodetool, we don't need to update it in the Medusa.