Merged PR 1110019: Nova - add medusa command to calculate backup size
This PR introduces a new command to calculate the size of backups.

When dealing with differential backups, simply summing up the sizes of each backup does not provide an accurate total size.

To address this, the command utilizes the manifest files associated with the given backups. The manifest file contains information such as the filename, size, and MD5 hash for each C* file. By iterating over these manifest files, the command stores each C* file name and its size in a map, eliminating any duplicate entries. Finally, the sizes of the files in the map are summed up to determine the overall backup size.

Additionally, certain backups need to be excluded from the size calculation. This is because we offer customers two free backups, which should not be included in the size calculation.